% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naive_rcpp_multiply}
\alias{naive_rcpp_multiply}
\title{Naive Matrix Multiplication (C++)}
\usage{
naive_rcpp_multiply(A, B)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}
}
\value{
The matrix product of A and B.
}
\description{
Computes the matrix product of two matrices using a simple, three-loop
naive algorithm in C++. This function is provided for educational and
benchmarking purposes.
}
