% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{strassen_hybrid}
\alias{strassen_hybrid}
\title{Hybrid Strassen's Matrix Multiplication (Sequential)}
\usage{
strassen_hybrid(A, B, threshold = 64L)
}
\arguments{
\item{A}{A numeric matrix.}

\item{B}{A numeric matrix.}

\item{threshold}{The matrix size at which the algorithm switches from Strassen's
recursion to a naive multiplication. Defaults to 64.}
}
\value{
The matrix product of A and B.
}
\description{
Computes the matrix product of two matrices using a sequential, hybrid
version of Strassen's algorithm. It uses a threshold to switch to a naive
algorithm for smaller sub-problems but does not use parallel processing.
}
